Handling internet disconnections. Minor other fixes.
