+++ CODIGO/app/activity_recognition/app/src/main/java/org/activityrecognition/user/LoginActivity.java
@@ -62,6 +62,16 @@ public class LoginActivity extends BaseActivity {
     }
 
     @Override
+    protected void disableActions() {
+
+    }
+
+    @Override
+    protected void updateView() {
+
+    }
+
+    @Override
     public void onResume() {
         super.onResume();
         signUpLink.setEnabled(true);
